+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # CHANGELOG
 
-## 0.0.1+1
+## 0.0.1
 
 * Initial release.
+
+## 0.0.2
+
+* Add language listeners
+* Update lang_tr.yaml configuration file